Actual fix format crashing on protected chars
The way we protected chars was by adding a \ before it, but it turns out this very behavior was the cause of the crashes.

Tweaking how we protect the strings (to protect from the String.format method) and removing over-protected char fixes most of the issues found, probably adding some regressions : more feedback should be given if the format method does not behave like it should.

Added tests to test the behavior of Utils.format. The tests are very basic and are only a quick indicator on the behavior that we expect (those tests wouldn't have passed before the fix).