Change Console method to force having a startImmediately boolean
This change is NOT backward compatible as it changes the forcefully Overridden initialize() method, forcefully needing a boolean value.

The previous implementation still work, it default to send initialize(false), but Console being an interface, you NEED to change your classes if those implements Console in any way.

Related to Vhoyon/Discord-Bot#159