Fix Logger problems and LoggerTest tests are now actually working
Logger had problems figuring out which class was it actually ran in and LoggerTest's only test was not actually testing anything.

Putting the work of building the message and handling it in a thread was more than simply wrapping them with a new thread constructor, and showed alot of problems that were actually affecting Logger.

This commit, as the summary suggests, fixes most of these issues while keeping threads for added optimization boosts.

To acheive these fixes, a new manager class was added : ThreadPool, which executes tasks in the future with workers that actively waits for new tasks to come. You can specify how many workers works in permanence using the ThreadPool(int) constructor, which gives addition performance boosts using a bigger parallelisation pool (more threads available to pickup tasks given to this ThreadPool object).