Remove actually problematic threading optimization in Logger
While in theory useful, this optimization was causing syncing issues and annoying delays, and therefore the Logger should not be optimized using the ThreadPool object.

The ThreadPool object is still used in Audit tho, as such Audit is not at all in the frontend and therefore is a really good optimization candidate using ThreadPool.