Refactor Settings related classes
- Setting -> SettingRepository
- SettingField -> Setting

This makes the notions of those two classes much more descriptive of their uses and makes further developments easier to understand.

THIS CHANGE IS A HUGE BACKWARD INCOMPATIBILITY however, and apps using this framework WILL NEED to adapt to this new naming scheme by changing the two classes above to the new class name specified. The ports might be a little more tedious a an old class has the same name has a new class (Setting), but their behavior are so different that any IDE should flag it instantly.