Separate commands with multiple calls to new Interface
This refines the LinkableCommand class while allowing to have further functionalities when using LinkableMultiCommand (you can even use both at the same time to get the actual Multi usage of LinkableMultiCommand).

When implementing LinkableMultiCommand, you do not need to Override getCall() anymore, but instead getCalls(), which expects a return of type String[]. All aliases must be in that array, the first element being the default call.

This commit changes alot of the structure of the commands, switching from Object handling to String return type, which means that projects using Vramework break if they upgrade to this version.