Tweak AbstractBotCommand to put a handler on the action method
This change makes it so that all BotCommands will need to Override the actions() method instead of the action() method to support vramework level features such as Bot Typing indicator.

This change introduces a new abstract method, and is therefore NOT backward compatible - old BotCommand commands will need to adapt to this new change (probably only requiring to add an "s" to the end of their "action()" method override).